Reorganised playbook integration test folders and improved playbook domain test
